Split IBDataProvider into its own module
