+++ bankroll/brokers/ibkr/__init__.py
@@ -1,3 +1,4 @@
-from .account import IBAccount, IBDataProvider, Settings, contract
+from .account import IBAccount, Settings
+from .marketdata import IBDataProvider, contract
 
 __all__ = ["IBAccount", "IBDataProvider", "Settings", "contract"]
